+++ internal/install/command_test.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -21,6 +22,7 @@ import (
 // 	testcobra.CheckCobraMetadata(t, Command)
 // 	testcobra.CheckCobraRequiredFlags(t, Command, []string{})
 // }
+
 func TestValidateProfile(t *testing.T) {
 	accountID := os.Getenv("NEW_RELIC_ACCOUNT_ID")
 	apiKey := os.Getenv("NEW_RELIC_API_KEY")
@@ -92,4 +94,30 @@ func initSegmentMockServer() *httptest.Server {
 		_, _ = w.Write([]byte(`[]`))
 	}))
 	return server
+}
+
+func TestProxyNetwork(t *testing.T) {
+
+	proxyConfig := struct {
+		HTTPSProxy string
+		HTTPProxy  string
+	}{
+		HTTPSProxy: "http://localhost:3128",
+		HTTPProxy:  "http://localhost:8080",
+	}
+
+	// Validate HTTPSProxy
+	if strings.HasPrefix(proxyConfig.HTTPSProxy, "http://") {
+		t.Log("New Relic CLI exclusively supports https proxy, not http for security reasons.")
+	} else if strings.HasPrefix(proxyConfig.HTTPSProxy, "https://") {
+		// Do nothing
+	} else {
+		t.Log("Invalid proxy provided")
+	}
+
+	// Validate HTTPProxy
+	if strings.HasPrefix(proxyConfig.HTTPProxy, "http://") {
+		t.Log("If you need to use a proxy, consider setting the HTTPS_PROXY environment variable, then try again. New Relic CLI exclusively supports https proxy.")
+	}
+
 }